Add note about building for something other than a PC, e.g. a RasPi
The Makefile is configured by settings on top, not by trying to completely determine the target from the kernel tree or other things, and some other options.  For a normal USB adapter on a PC, just type "make", for anything else, it needs the options changed.  The added section explains it.